ProyectoP1/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
click==8.1.8
colorama==0.4.6
distlib==0.3.9
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
filelock==3.18.0
gunicorn==23.0.0
h11==0.14.0
packaging==24.2
platformdirs==4.3.7
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2
uvicorn==0.34.1
virtualenv==20.30.0
whitenoise==6.9.0
